+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 * 作成
 ## 0.1.0
 * エラーをObject型からジェネリクスへ変更
+## 0.2.0
+* fetcherとinfinite_fetcherの引数を名前付き引数に変更
+* infinite_fetcherの引数の_fetchNewをnullableに変更